# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IA Calorias API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type UserTier = (typeof UserTier)[keyof typeof UserTier];

export const UserTier = {
  free: "free",
  limited: "limited",
  unlimited: "unlimited",
} as const;

export interface User {
  sessionId: string;
  tier: UserTier;
  analysisCount: number;
  analysisLimit?: number | null;
  trialRemaining: number;
}

export interface MacroNutrients {
  protein: number;
  carbs: number;
  fat: number;
}

export interface AnalysisResult {
  id: string;
  sessionId: string;
  dishName: string;
  calories: number;
  macros: MacroNutrients;
  imageUrl?: string | null;
  createdAt: string;
}

export interface PaywallResponse {
  error: string;
  message?: string;
  requiresUpgrade: boolean;
  trialUsed: number;
  trialLimit: number;
}

export type SubscriptionStatusTier =
  (typeof SubscriptionStatusTier)[keyof typeof SubscriptionStatusTier];

export const SubscriptionStatusTier = {
  free: "free",
  limited: "limited",
  unlimited: "unlimited",
} as const;

export interface SubscriptionStatus {
  sessionId: string;
  tier: SubscriptionStatusTier;
  analysisCount: number;
  analysisLimit?: number | null;
  trialRemaining: number;
  stripeCustomerId?: string | null;
  stripeSubscriptionId?: string | null;
  currentPeriodEnd?: string | null;
}

export type CheckoutRequestPlan =
  (typeof CheckoutRequestPlan)[keyof typeof CheckoutRequestPlan];

export const CheckoutRequestPlan = {
  limited: "limited",
  unlimited: "unlimited",
} as const;

export interface CheckoutRequest {
  sessionId: string;
  plan: CheckoutRequestPlan;
}

export interface CheckoutResponse {
  url: string;
}

export type AnalyzeFoodBody = {
  image: Blob;
  sessionId: string;
};

export type GetAnalysisHistoryParams = {
  sessionId: string;
};

export type GetSubscriptionStatusParams = {
  sessionId: string;
};

export type StripeWebhookBody = { [key: string]: unknown };