# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IA Calorias API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: user
    description: User operations
  - name: analysis
    description: Food analysis operations
  - name: subscription
    description: Subscription and payment operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /user/me:
    get:
      operationId: getMe
      tags: [user]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analysis:
    post:
      operationId: analyzeFood
      tags: [analysis]
      summary: Analyze food image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                sessionId:
                  type: string
              required:
                - image
                - sessionId
      responses:
        "200":
          description: Food analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "402":
          description: Payment required - trial limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaywallResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analysis/history:
    get:
      operationId: getAnalysisHistory
      tags: [analysis]
      summary: Get analysis history for session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalysisResult"

  /subscription/status:
    get:
      operationId: getSubscriptionStatus
      tags: [subscription]
      summary: Get subscription status for session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"

  /subscription/checkout:
    post:
      operationId: createCheckoutSession
      tags: [subscription]
      summary: Create Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /subscription/webhook:
    post:
      operationId: stripeWebhook
      tags: [subscription]
      summary: Stripe webhook handler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    User:
      type: object
      properties:
        sessionId:
          type: string
        tier:
          type: string
          enum: [free, limited, unlimited]
        analysisCount:
          type: integer
        analysisLimit:
          type: integer
          nullable: true
        trialRemaining:
          type: integer
      required:
        - sessionId
        - tier
        - analysisCount
        - trialRemaining

    MacroNutrients:
      type: object
      properties:
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
      required:
        - protein
        - carbs
        - fat

    AnalysisResult:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        dishName:
          type: string
        calories:
          type: integer
        macros:
          $ref: "#/components/schemas/MacroNutrients"
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - dishName
        - calories
        - macros
        - createdAt

    PaywallResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        requiresUpgrade:
          type: boolean
        trialUsed:
          type: integer
        trialLimit:
          type: integer
      required:
        - error
        - requiresUpgrade
        - trialUsed
        - trialLimit

    SubscriptionStatus:
      type: object
      properties:
        sessionId:
          type: string
        tier:
          type: string
          enum: [free, limited, unlimited]
        analysisCount:
          type: integer
        analysisLimit:
          type: integer
          nullable: true
        trialRemaining:
          type: integer
        stripeCustomerId:
          type: string
          nullable: true
        stripeSubscriptionId:
          type: string
          nullable: true
        currentPeriodEnd:
          type: string
          format: date-time
          nullable: true
      required:
        - sessionId
        - tier
        - analysisCount
        - trialRemaining

    CheckoutRequest:
      type: object
      properties:
        sessionId:
          type: string
        plan:
          type: string
          enum: [limited, unlimited]
      required:
        - sessionId
        - plan

    CheckoutResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url