# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IA Calorias API
 * OpenAPI spec version: 0.1.0
 */
import type { MacroNutrients } from "./macroNutrients";

export interface AnalysisResult {
  id: string;
  sessionId: string;
  dishName: string;
  calories: number;
  macros: MacroNutrients;
  imageUrl?: string | null;
  createdAt: Date;
}