# lib/api-zod/src/generated/types/checkoutRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IA Calorias API
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutRequestPlan } from "./checkoutRequestPlan";

export interface CheckoutRequest {
  sessionId: string;
  plan: CheckoutRequestPlan;
}