# lib/api-zod/src/generated/types/macroNutrients.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IA Calorias API
 * OpenAPI spec version: 0.1.0
 */

export interface MacroNutrients {
  protein: number;
  carbs: number;
  fat: number;
}