# lib/api-zod/src/generated/types/paywallResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IA Calorias API
 * OpenAPI spec version: 0.1.0
 */

export interface PaywallResponse {
  error: string;
  message?: string;
  requiresUpgrade: boolean;
  trialUsed: number;
  trialLimit: number;
}