# lib/api-zod/src/generated/types/subscriptionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IA Calorias API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionStatusTier } from "./subscriptionStatusTier";

export interface SubscriptionStatus {
  sessionId: string;
  tier: SubscriptionStatusTier;
  analysisCount: number;
  analysisLimit?: number | null;
  trialRemaining: number;
  stripeCustomerId?: string | null;
  stripeSubscriptionId?: string | null;
  currentPeriodEnd?: Date | null;
}