# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IA Calorias API
 * OpenAPI spec version: 0.1.0
 */
import type { UserTier } from "./userTier";

export interface User {
  sessionId: string;
  tier: UserTier;
  analysisCount: number;
  analysisLimit?: number | null;
  trialRemaining: number;
}